+++ test/libsolidity/SolidityEndToEndTest.cpp
@@ -7916,7 +7916,7 @@ BOOST_AUTO_TEST_CASE(string_tuples)
 				return (h(), "def");
 			}
 			function h() public returns (string) {
-				return ("abc",);
+				return ("abc");
 			}
 		}
 	)";